requirements.txt
python-pptx>=0.6.23
pymupdf>=1.24.0
Pillow>=10.0.0
python-docx>=1.1.0
